:sparkles: feat: adiciona middleware de autenticação e validação de dono da oferta para exclusão e edição

+++ src/controllers/skill/deleteSkill.js
@@ -2,11 +2,20 @@ import prisma from "../../PrismaCliente.js";
 
 const deleteSkill = async (req, res) => {
     const { id } = req.params;
+    const loggedPersonId = req.loggedPersonId;
+
     try {
         const skill = await prisma.conhecimento.findUnique({ where: { id: Number(id) } });
 
         if (!skill) {
             return res.status(404).json("Conhecimento não encontrado.");
+        };
+
+        if (skill.pessoa_id !== loggedPersonId) {
+
+            return res.status(403).json({
+                mensagem: "Você não tem permissão para excluir uma oferta que não é sua."
+            });
         };
 
         await prisma.conhecimento.delete({ where: { id: Number(id) } });
